Updated data paths to match test event. Timestamp for random ticket id. Refactored resuable methods
